+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 1. (SidePanel)管理页面回到顶部按钮。
 2. (SidePanel)管理页面搜索表格新增列：最低薪资，最高薪资，几薪，学历。
 3. (SidePanel)管理页面搜索表格新增排序列：首次浏览日期，发布日期，最低薪资，最高薪资，几薪，学历。
+4. (ContentScript)显示职位发布时间与自动排序（支持jobsdb-hk）。
 
 ### ✏️ Changed
 
@@ -20,7 +21,7 @@
 
 ### ⭐ Added
 
-1. (ContentScript)显示职位发布时间（支持BOSS,51Job,智联,拉钩）。
+1. (ContentScript)显示职位发布时间与自动排序（支持BOSS,51Job,智联,拉钩）。
 2. (ContentScript)显示外包公司和培训机构。
 3. (ContentScript)职位评论。
 4. (SidePanel)新增本地显示职位初次浏览时间，历史浏览次数。